[ADMIN] SR - Move methods from controller to component
