+++ admin/app/routes/authenticated/target-profiles/target-profile/organizations.js
@@ -1,5 +1,4 @@
 import Route from '@ember/routing/route';
-import { action } from '@ember/object';
 
 export default class TargetProfileOrganizationsRoute extends Route {
   queryParams = {
@@ -41,12 +40,6 @@ export default class TargetProfileOrganizationsRoute extends Route {
       controller.name = null;
       controller.type = null;
       controller.externalId = null;
-      controller.organizationsToAttach = [];
     }
-  }
-
-  @action
-  refreshModel() {
-    this.refresh();
   }
 }
