+++ admin/app/templates/authenticated/target-profiles/target-profile/organizations.hbs
@@ -2,14 +2,11 @@
 
 <TargetProfiles::Organizations
   @organizations={{@model.organizations}}
+  @targetProfile={{@model.targetProfile}}
   @id={{this.id}}
   @name={{this.name}}
   @type={{this.type}}
   @externalId={{this.externalId}}
   @triggerFiltering={{this.triggerFiltering}}
   @goToOrganizationPage={{this.goToOrganizationPage}}
-  @organizationsToAttach={{this.organizationsToAttach}}
-  @attachOrganizations={{this.attachOrganizations}}
-  @existingTargetProfile={{this.existingTargetProfile}}
-  @attachOrganizationsFromExistingTargetProfile={{this.attachOrganizationsFromExistingTargetProfile}}
 />
