+++ admin/tests/integration/components/target-profiles/organizations_test.js
@@ -5,15 +5,12 @@ import hbs from 'htmlbars-inline-precompile';
 import EmberObject from '@ember/object';
 import fillInByLabel from '../../../helpers/extended-ember-test-helpers/fill-in-by-label';
 import clickByLabel from '../../../helpers/extended-ember-test-helpers/click-by-label';
-import sinon from 'sinon';
 
 module('Integration | Component | TargetProfiles::Organizations', function(hooks) {
 
   setupRenderingTest(hooks);
 
   hooks.beforeEach(function() {
-    this.attachOrganizations = () => {};
-    this.attachOrganizationsFromExistingTargetProfile = () => {};
     this.triggerFiltering = () => {};
     this.goToOrganizationPage = () => {};
   });
@@ -27,7 +24,7 @@ module('Integration | Component | TargetProfiles::Organizations', function(hooks
     this.organizations = organizations;
 
     // when
-    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @attachOrganizations={{this.attachOrganizations}} @attachOrganizationsFromExistingTargetProfile={{this.attachOrganizationsFromExistingTargetProfile}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}} />`);
+    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}} />`);
 
     // then
     assert.contains('Orga1');
@@ -42,7 +39,7 @@ module('Integration | Component | TargetProfiles::Organizations', function(hooks
     this.organizations = organizations;
 
     // when
-    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @attachOrganizations={{this.attachOrganizations}} @attachOrganizationsFromExistingTargetProfile={{this.attachOrganizationsFromExistingTargetProfile}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}} />`);
+    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}} />`);
 
     // then
     assert.contains('Orga1');
@@ -57,7 +54,7 @@ module('Integration | Component | TargetProfiles::Organizations', function(hooks
     this.organizations = [];
 
     // when
-    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @organizationsToAttach={{this.organizationsToAttach}} @attachOrganizations={{this.attachOrganizations}} @attachOrganizationsFromExistingTargetProfile={{this.attachOrganizationsFromExistingTargetProfile}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
+    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
 
     assert.contains('Aucun résultat');
   });
@@ -65,38 +62,24 @@ module('Integration | Component | TargetProfiles::Organizations', function(hooks
   test('it should display a field to attach organizations', async function(assert) {
     // given
     this.organizations = [];
-    this.organizationsToAttach = null;
-    const attachOrganizations = sinon.stub();
-    const attachOrganizationsWrapper = function(e) {
-      e.preventDefault();
-      attachOrganizations();
-    };
-    this.attachOrganizations = attachOrganizationsWrapper;
 
     // when
-    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @organizationsToAttach={{this.organizationsToAttach}} @attachOrganizations={{this.attachOrganizations}} @attachOrganizationsFromExistingTargetProfile={{this.attachOrganizationsFromExistingTargetProfile}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
+    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
     await fillInByLabel('Rattacher une ou plusieurs organisation(s)', '1, 2');
     await clickByLabel('Valider le rattachement');
 
-    assert.ok(attachOrganizations.called);
+    assert.dom('[placeholder="1, 2"]').hasValue('1, 2');
   });
 
   test('it should display a field to attach organizations from an existing target profile', async function(assert) {
     // given
     this.organizations = [];
-    this.existingTargetProfile = null;
-    const attachOrganizationsFromExistingTargetProfile = sinon.stub();
-    const attachOrganizationsFromExistingTargetProfileWrapper = function(e) {
-      e.preventDefault();
-      attachOrganizationsFromExistingTargetProfile();
-    };
-    this.attachOrganizationsFromExistingTargetProfile = attachOrganizationsFromExistingTargetProfileWrapper;
 
     // when
-    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @existingTargetProfile={{this.existingTargetProfile}} @attachOrganizations={{this.attachOrganizations}} @attachOrganizationsFromExistingTargetProfile={{attachOrganizationsFromExistingTargetProfile}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
+    await render(hbs`<TargetProfiles::Organizations @organizations={{this.organizations}} @goToOrganizationPage={{this.goToOrganizationPage}} @triggerFiltering={{this.triggerFiltering}}/>`);
     await fillInByLabel('Rattacher les organisations d\'un profil cible existant', 1);
     await clickByLabel('Valider le rattachement à partir de ce profil cible');
 
-    assert.ok(attachOrganizationsFromExistingTargetProfile.called);
+    assert.dom('[placeholder="1135"]').hasValue('1');
   });
 });
